requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
lxml==4.7.1
oauthlib==3.1.1
python-docx==0.8.11
requests==2.27.1
requests-oauthlib==1.3.0
soupsieve==2.3.1
tweepy==4.4.0
urllib3==1.26.7
